Change graph fields types from int to size_t, Add wotking test scripts

+++ Course 2/Algorithms_and_Data_Structures/Labs2-3/src/algorithms/algorithms.h
@@ -4,8 +4,8 @@
 
 using namespace std;
 
-list<Vertex*> Dijkstra(Graph* graph, int startId, int destinationId);
-list<Vertex*> BellmanFord(Graph* graph, int startId, int destinationId);
+list<Vertex*> Dijkstra(Graph* graph, size_t startId, size_t destinationId);
+list<Vertex*> BellmanFord(Graph* graph, size_t startId, size_t destinationId);
 
 Graph* Prim(Graph* graph);
 Graph* Kruskal(Graph* graph);
